size_t to int

+++ config.h
@@ -1,3 +1,4 @@
 static const int samplerate = 44100;
+
 // 5 guesses in one second
-static const size_t buflen = 8820;
+static const int buflen = 8820;
